+++ pyoffers/models/core.py
@@ -20,21 +20,8 @@ class Model:
     def __getitem__(self, item):
         return self._data[item]
 
-    def __setitem__(self, key, value):
-        self._data[key] = value
-
     def __eq__(self, other):
         return isinstance(other, self.__class__) and self['id'] == other['id']
-
-    @property
-    def instance_data(self):
-        """
-        All data except for private and non model fields.
-        """
-        return {
-            key: value for key, value in self._data.items()
-            if not key.startswith('_') and key not in self.non_model_fields
-        }
 
 
 class ModelManager: